Fix the Unpack qualified typing rule
I think I had added these omega during the call. I got confused
there. Even ignoring that they are in the wrong place (if there needs
to be omega, they should be in the conclusion), we absolutely don't
want them, otherwise we would sort of force the linear constraint `Q`
to be unrestricted, which is not useful for our case.

Here we really want to consume `e1` once, the value packed in
`e1` (besides the constraint) is unrestricted, and it is reflected in
the `omega` index for the `x` variable. This is the only omega we
need.

We could add a scaling factor `pi` like for `case`, but it's not
useful for our applications.