Some progress on the new example in the intro
Talking about the full-blown PArray API is difficult here. Instead,
I'll use UArray (as in, an array whose elements are unrestricted things) here
that is more comparable in power to 'MArray' from linear-base. Then in
S4 we can define introduce PArray and define UArray as a synonym for
PArray (AtomRef a)

With UArray it would be possible to have a simple 'swap' implementation,
but this new example is more focused on 'free' which is now a stronger
and more concise claim about safety. Instead, I'll define 'swap' for
PArray later