Use entr instead of inotify-wait for the continuous builder
It turns out to be much more robust.

I also removed the internal call to `nix-shell` for snappier
rebuilds. The consequence is that one needs to manually reload if the
`shell.nix` file is modified. There are probably ways to handle
`shell.nix` as well, using an outer loop and restarting the
process. But I don't want to figure this out today.