Remove the “linear lets are strict” footnote
This remark only applies to destructuring lets. Of which the only use
in the paper is in the first example of the introduction. Which we say
we want to replace. The rest of the time, we use let⬜, which is
necessarily strict as it's a form of `case` (we don't speak about it,
but it should be obvious to anybody who knows enough to ask the
question). This remark is therefore not technically relevant for the
paper.